+++ CHANGELOG.md
@@ -34,6 +34,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - LASRV007 - 27/03/2023 - Add the ProcessDeletion class to kill process.
 - LASRV008 - 28/03/2003 - Add the EventInformation class to get information of event.
 - LASRV008 - 28/03/2003 - Add the EventInsertion class to insert event.
+- LASRV008 - 28/03/2003 - Add the EventModificator class to modify event.
+
 
 ### Changed
 - LASRV001 - 21/03/2023 - Refactor code in the temperature detectors.